+++ Sources/RiviumTrace/Utils/RiviumTraceVersion.swift
@@ -2,5 +2,5 @@ import Foundation
 
 /// Central SDK version constant
 public enum RiviumTraceSDK {
-    public static let version = "2.0.0"
+    public static let version = "0.2.0"
 }